+++ src/seeps/mod.ts
@@ -1,2 +1,2 @@
-export * from './or-seep.js';
+export * from './or.seep.js';
 export * from './seep.js';
